Quadratic Voting: Remove maxVoterWeight after discussions with Solana Labs about how to track maxVoterWeight.
A summary of the discussion:

This is an issue that also applies to VSR DAOs, but that it is essentially a UX issue, and relates to the decision of "how to set a quorum correctly"?

Essentially - in "normal" DAOs, you set the quorum as a percentage of the total vote power, where vote power = token supply.

With QV DAOs, the total vote power is a function of total supply and distribution.

If the distribution is wide and egalitarian (uniform distribution of tokens), the total power is higher than if the distribution is narrow and centralised in a few whales (long-tail distribution).

So this leads to two facts:
- calculating max weight is complicated and changes over time - we could do it, but it would require changes in the core program as mentioned.
- setting the quorum as a percentage of the max weight is not that useful, since, depending on the distribution, that percentage may be reachable by one or two whales (note, this is even worse in token-voting DAOs.

As a result, we decided it is currently not necessary to calculate an accurate max weight, and quorum will be determined based on the token supply, with potentially some UI help to explain how that is affected by QV.